+++ src/acpype/__init__.py
@@ -54,4 +54,4 @@ Alan Silva, D.Sc. <alanwilter _at_ gmail _dot_ com>
 # from https://packaging.python.org/guides/single-sourcing-package-version/
 # using option 2
 # updated automatically via pre-commit git-hook
-__version__ = "2026.9.4"
+__version__ = "2026.9.5"